chore: reorder fakes
