+++ src/Arius.Core.Tests/Shared/FileTree/FileTreeBuilderTests.cs
@@ -5,7 +5,7 @@ using Arius.Core.Shared.FileTree;
 using Arius.Core.Shared.Hashes;
 using Arius.Core.Shared.Storage;
 using Arius.Core.Tests.Fakes;
-using System.Collections.Concurrent;
+using Arius.Core.Tests.Shared.FileTree.Fakes;
 
 namespace Arius.Core.Tests.Shared.FileTree;
 
@@ -25,160 +25,6 @@ public class FileTreeBuilderTests
             await writer.AppendFileEntryAsync(file.Path, file.Hash, file.Created, file.Modified);
 
         return (session, session.StagingRoot);
-    }
-
-    private sealed class BlockingFileTreeUploadBlobContainerService : IBlobContainerService
-    {
-        private readonly TaskCompletionSource<bool> _allowUploads = new(TaskCreationOptions.RunContinuationsAsynchronously);
-        private readonly TaskCompletionSource<bool> _twoUploadsStarted = new(TaskCreationOptions.RunContinuationsAsynchronously);
-        private int _startedUploads;
-
-        public ConcurrentDictionary<string, byte> Uploaded { get; } = new(StringComparer.Ordinal);
-
-        public Task CreateContainerIfNotExistsAsync(CancellationToken cancellationToken = default) => Task.CompletedTask;
-
-        public async Task UploadAsync(string blobName, Stream content, IReadOnlyDictionary<string, string> metadata, BlobTier tier, string? contentType = null, bool overwrite = false, CancellationToken cancellationToken = default)
-        {
-            if (blobName.StartsWith(BlobPaths.FileTrees, StringComparison.Ordinal))
-            {
-                if (Interlocked.Increment(ref _startedUploads) >= 2)
-                    _twoUploadsStarted.TrySetResult(true);
-
-                await _allowUploads.Task.WaitAsync(cancellationToken);
-            }
-
-            Uploaded.TryAdd(blobName, 0);
-        }
-
-        public Task<Stream> OpenWriteAsync(string blobName, string? contentType = null, CancellationToken cancellationToken = default) =>
-            Task.FromResult<Stream>(new MemoryStream());
-
-        public Task<Stream> DownloadAsync(string blobName, CancellationToken cancellationToken = default) =>
-            Task.FromResult<Stream>(new MemoryStream());
-
-        public Task<BlobMetadata> GetMetadataAsync(string blobName, CancellationToken cancellationToken = default) =>
-            Task.FromResult(new BlobMetadata { Exists = false });
-
-        public IAsyncEnumerable<string> ListAsync(string prefix, CancellationToken cancellationToken = default) =>
-            AsyncEnumerable.Empty<string>();
-
-        public Task SetMetadataAsync(string blobName, IReadOnlyDictionary<string, string> metadata, CancellationToken cancellationToken = default) => Task.CompletedTask;
-        public Task SetTierAsync(string blobName, BlobTier tier, CancellationToken cancellationToken = default) => Task.CompletedTask;
-        public Task CopyAsync(string sourceBlobName, string destinationBlobName, BlobTier destinationTier, RehydratePriority? rehydratePriority = null, CancellationToken cancellationToken = default) => Task.CompletedTask;
-        public Task DeleteAsync(string blobName, CancellationToken cancellationToken = default) => Task.CompletedTask;
-
-        public async Task<bool> WaitForTwoUploadsAsync(TimeSpan timeout)
-        {
-            using var timeoutCts = new CancellationTokenSource(timeout);
-            try
-            {
-                await _twoUploadsStarted.Task.WaitAsync(timeoutCts.Token);
-                return true;
-            }
-            catch (OperationCanceledException)
-            {
-                return false;
-            }
-        }
-
-        public void AllowUploads() => _allowUploads.TrySetResult(true);
-    }
-
-    private sealed class CountingFileTreeUploadBlobContainerService : IBlobContainerService
-    {
-        private readonly TaskCompletionSource<bool> _releaseUploads = new(TaskCreationOptions.RunContinuationsAsynchronously);
-        private int _currentUploads;
-        private int _maxConcurrentUploads;
-
-        public int MaxConcurrentUploads => _maxConcurrentUploads;
-
-        public Task CreateContainerIfNotExistsAsync(CancellationToken cancellationToken = default) => Task.CompletedTask;
-
-        public async Task UploadAsync(string blobName, Stream content, IReadOnlyDictionary<string, string> metadata, BlobTier tier, string? contentType = null, bool overwrite = false, CancellationToken cancellationToken = default)
-        {
-            if (!blobName.StartsWith(BlobPaths.FileTrees, StringComparison.Ordinal))
-                return;
-
-            var currentUploads = Interlocked.Increment(ref _currentUploads);
-            UpdateMaxConcurrentUploads(currentUploads);
-
-            try
-            {
-                await _releaseUploads.Task.WaitAsync(cancellationToken);
-            }
-            finally
-            {
-                Interlocked.Decrement(ref _currentUploads);
-            }
-        }
-
-        public Task<Stream> OpenWriteAsync(string blobName, string? contentType = null, CancellationToken cancellationToken = default) =>
-            Task.FromResult<Stream>(new MemoryStream());
-
-        public Task<Stream> DownloadAsync(string blobName, CancellationToken cancellationToken = default) =>
-            Task.FromResult<Stream>(new MemoryStream());
-
-        public Task<BlobMetadata> GetMetadataAsync(string blobName, CancellationToken cancellationToken = default) =>
-            Task.FromResult(new BlobMetadata { Exists = false });
-
-        public IAsyncEnumerable<string> ListAsync(string prefix, CancellationToken cancellationToken = default) =>
-            AsyncEnumerable.Empty<string>();
-
-        public Task SetMetadataAsync(string blobName, IReadOnlyDictionary<string, string> metadata, CancellationToken cancellationToken = default) => Task.CompletedTask;
-        public Task SetTierAsync(string blobName, BlobTier tier, CancellationToken cancellationToken = default) => Task.CompletedTask;
-        public Task CopyAsync(string sourceBlobName, string destinationBlobName, BlobTier destinationTier, RehydratePriority? rehydratePriority = null, CancellationToken cancellationToken = default) => Task.CompletedTask;
-        public Task DeleteAsync(string blobName, CancellationToken cancellationToken = default) => Task.CompletedTask;
-
-        public void AllowUploads() => _releaseUploads.TrySetResult(true);
-
-        private void UpdateMaxConcurrentUploads(int currentUploads)
-        {
-            while (true)
-            {
-                var observedMax = _maxConcurrentUploads;
-                if (currentUploads <= observedMax)
-                    return;
-
-                if (Interlocked.CompareExchange(ref _maxConcurrentUploads, currentUploads, observedMax) == observedMax)
-                    return;
-            }
-        }
-    }
-
-    private sealed class FaultingAndBlockingFileTreeUploadBlobContainerService : IBlobContainerService
-    {
-        private readonly TaskCompletionSource<bool> _blockedUploads = new(TaskCreationOptions.RunContinuationsAsynchronously);
-        private int _fileTreeUploads;
-
-        public Task CreateContainerIfNotExistsAsync(CancellationToken cancellationToken = default) => Task.CompletedTask;
-
-        public async Task UploadAsync(string blobName, Stream content, IReadOnlyDictionary<string, string> metadata, BlobTier tier, string? contentType = null, bool overwrite = false, CancellationToken cancellationToken = default)
-        {
-            if (!blobName.StartsWith(BlobPaths.FileTrees, StringComparison.Ordinal))
-                return;
-
-            if (Interlocked.Increment(ref _fileTreeUploads) == 1)
-                throw new InvalidOperationException("Simulated filetree upload failure.");
-
-            await _blockedUploads.Task.WaitAsync(cancellationToken);
-        }
-
-        public Task<Stream> OpenWriteAsync(string blobName, string? contentType = null, CancellationToken cancellationToken = default) =>
-            Task.FromResult<Stream>(new MemoryStream());
-
-        public Task<Stream> DownloadAsync(string blobName, CancellationToken cancellationToken = default) =>
-            Task.FromResult<Stream>(new MemoryStream());
-
-        public Task<BlobMetadata> GetMetadataAsync(string blobName, CancellationToken cancellationToken = default) =>
-            Task.FromResult(new BlobMetadata { Exists = false });
-
-        public IAsyncEnumerable<string> ListAsync(string prefix, CancellationToken cancellationToken = default) =>
-            AsyncEnumerable.Empty<string>();
-
-        public Task SetMetadataAsync(string blobName, IReadOnlyDictionary<string, string> metadata, CancellationToken cancellationToken = default) => Task.CompletedTask;
-        public Task SetTierAsync(string blobName, BlobTier tier, CancellationToken cancellationToken = default) => Task.CompletedTask;
-        public Task CopyAsync(string sourceBlobName, string destinationBlobName, BlobTier destinationTier, RehydratePriority? rehydratePriority = null, CancellationToken cancellationToken = default) => Task.CompletedTask;
-        public Task DeleteAsync(string blobName, CancellationToken cancellationToken = default) => Task.CompletedTask;
     }
 
     private static FileTreeBuilder CreateBuilder(
